fix test with mocking due to su
The related client has a `su` scope which provides the right to every scope.
The test assumes the scope `wibble` should cause a error due to not explicitly
owning the scope, but a recent fix in Persona fixed that issue.

The new mocking logic causes the call to fail as expected.